+++ android/src/main/kotlin/io/intheloup/geolocation/GeolocationPlugin.kt
@@ -16,7 +16,7 @@ class GeolocationPlugin(val registrar: Registrar) {
 
     init {
         registrar.addRequestPermissionsResultListener(locationClient.permissionResultListener)
-
+        registrar.addActivityResultListener(locationClient.activityResultListener);
         registrar.activity().application.registerActivityLifecycleCallbacks(object : Application.ActivityLifecycleCallbacks {
             override fun onActivityPaused(activity: Activity?) {
                 locationClient.pause()
@@ -59,6 +59,7 @@ class GeolocationPlugin(val registrar: Registrar) {
     }
 
     object Intents {
-        const val LocationPermissionRequestId = 138978923
+        const val LocationPermissionRequestId = 12234
+        const val EnableLocationSettingsRequestId = 12237
     }
 }