+++ android/src/main/kotlin/io/intheloup/geolocation/data/LocationData.kt
@@ -7,13 +7,15 @@ import android.location.Location
 
 class LocationData(val latitude: Double,
                    val longitude: Double,
-                   val altitude: Double
+                   val altitude: Double,
+                   val isMocked: Boolean
 ) {
     companion object {
         fun from(location: Location) = LocationData(
                 latitude = location.latitude,
                 longitude = location.longitude,
-                altitude = location.altitude
+                altitude = location.altitude,
+                isMocked = location.isFromMockProvider()
         )
     }
 }