+++ android/src/main/kotlin/io/intheloup/geolocation/location/LocationHelper.kt
@@ -8,7 +8,7 @@ import android.content.Context
 import android.content.pm.PackageManager
 import android.os.Build
 import android.provider.Settings
-import android.support.v4.content.ContextCompat
+import androidx.core.content.ContextCompat
 import com.google.android.gms.location.LocationRequest
 import io.intheloup.geolocation.data.Permission