+++ lib/facet_android/location.dart
@@ -44,4 +44,13 @@ class LocationOptionsAndroid {
 
   /// Always `1` in single location request
   final int numUpdates;
+
+  Map toJson() => {
+        'interval': interval,
+        'fastestInterval': fastestInterval,
+        'expirationTime': expirationTime,
+        'expirationDuration': expirationDuration,
+        'maxWaitTime': maxWaitTime,
+        'numUpdates': numUpdates
+      };
 }